cleanup - move port package into application package
