+++ configuration/src/main/kotlin/github/com/hukuta94/delivery/configuration/core/application/usecase/UseCaseConfiguration.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.configuration.core.application.usecase
 
 import github.com.hukuta94.delivery.core.application.usecase.event.impl.SaveIntegrationEventUseCaseImpl
-import github.com.hukuta94.delivery.core.port.repository.event.InboxEventRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.event.InboxEventRepositoryPort
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
