+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/event/domain/handler/OrderAssignedDomainEventHandler.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.core.application.event.domain.handler
 
 import github.com.hukuta94.delivery.core.domain.order.OrderAssignedDomainEvent
-import github.com.hukuta94.delivery.core.port.BusProducer
+import github.com.hukuta94.delivery.core.application.port.BusProducer
 
 class OrderAssignedDomainEventHandler(
     private val busProducer: BusProducer,
