+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/event/domain/handler/OrderCompletedDomainEventHandler.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.core.application.event.domain.handler
 
 import github.com.hukuta94.delivery.core.domain.order.OrderCompletedDomainEvent
-import github.com.hukuta94.delivery.core.port.BusProducer
+import github.com.hukuta94.delivery.core.application.port.BusProducer
 
 class OrderCompletedDomainEventHandler(
     private val busProducer: BusProducer,
