+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/port/GetLocationPort.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.port
+package github.com.hukuta94.delivery.core.application.port
 
 import github.com.hukuta94.delivery.core.domain.common.Location
