+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/port/repository/domain/AggregateRepository.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.core.port.repository.domain
+package github.com.hukuta94.delivery.core.application.port.repository.domain
 
 import github.com.hukuta94.delivery.core.domain.Aggregate
 import java.util.*
