+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/usecase/event/impl/SaveIntegrationEventUseCaseImpl.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.core.application.usecase.event.impl
 
 import github.com.hukuta94.delivery.core.application.usecase.event.SaveIntegrationEventUseCase
 import github.com.hukuta94.delivery.core.domain.IntegrationEvent
-import github.com.hukuta94.delivery.core.port.repository.event.InboxEventRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.event.InboxEventRepositoryPort
 
 class SaveIntegrationEventUseCaseImpl(
     private val inboxEventRepositoryPort: InboxEventRepositoryPort,
