+++ core/impl/src/main/kotlin/github/com/hukuta94/delivery/core/application/usecase/order/impl/AssignCourierToOrderUseCaseImpl.kt
@@ -2,9 +2,9 @@ package github.com.hukuta94.delivery.core.application.usecase.order.impl
 
 import github.com.hukuta94.delivery.core.application.usecase.order.AssignCourierToOrderUseCase
 import github.com.hukuta94.delivery.core.domain.service.DispatchService
-import github.com.hukuta94.delivery.core.port.repository.domain.CourierRepositoryPort
-import github.com.hukuta94.delivery.core.port.repository.domain.OrderRepositoryPort
-import github.com.hukuta94.delivery.core.port.repository.UnitOfWorkPort
+import github.com.hukuta94.delivery.core.application.port.repository.domain.CourierRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.domain.OrderRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.UnitOfWorkPort
 import org.slf4j.LoggerFactory
 
 class AssignCourierToOrderUseCaseImpl(
