+++ core/tests/src/test/kotlin/github/com/hukuta94/delivery/core/domain/common/MoneyTest.kt
@@ -1,6 +1,6 @@
 package github.com.hukuta94.delivery.core.domain.common
 
-import github.com.hukuta94.delivery.core.port.Exchange
+import github.com.hukuta94.delivery.core.application.port.Exchange
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldNotBe
 import org.junit.jupiter.api.Test
