+++ infrastructure/src/main/kotlin/github/com/hukuta94/delivery/infrastructure/adapter/grpc/GeoServiceGrpcClient.kt
@@ -1,7 +1,7 @@
 package github.com.hukuta94.delivery.infrastructure.adapter.grpc
 
 import github.com.hukuta94.delivery.core.domain.common.Location
-import github.com.hukuta94.delivery.core.port.GetLocationPort
+import github.com.hukuta94.delivery.core.application.port.GetLocationPort
 import github.com.hukuta94.delivery.infrastructure.adapter.grpc.GeoGrpc.newBlockingStub
 import io.grpc.ManagedChannel
