+++ infrastructure/src/main/kotlin/github/com/hukuta94/delivery/infrastructure/adapter/inmemory/OrderInMemoryRepository.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.infrastructure.adapter.inmemory
 
 import github.com.hukuta94.delivery.core.domain.order.Order
 import github.com.hukuta94.delivery.core.domain.order.OrderStatus
-import github.com.hukuta94.delivery.core.port.repository.domain.OrderRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.domain.OrderRepositoryPort
 import java.util.*
 
 class OrderInMemoryRepository : OrderRepositoryPort {
