+++ infrastructure/src/main/kotlin/github/com/hukuta94/delivery/infrastructure/adapter/orm/repository/domain/OrmOrderRepositoryAdapter.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.domai
 
 import github.com.hukuta94.delivery.core.domain.order.Order
 import github.com.hukuta94.delivery.core.domain.order.OrderStatus
-import github.com.hukuta94.delivery.core.port.repository.domain.OrderRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.domain.OrderRepositoryPort
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.model.entity.OrderJpaEntity
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.event.OrmOutboxEventRepositoryAdapter
 import java.util.*
