+++ infrastructure/src/main/kotlin/github/com/hukuta94/delivery/infrastructure/adapter/orm/repository/event/OrmInboxEventRepositoryAdapter.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.event
 
 import github.com.hukuta94.delivery.core.domain.IntegrationEvent
 import github.com.hukuta94.delivery.core.application.event.integration.IntegrationEventSerializer
-import github.com.hukuta94.delivery.core.port.repository.event.InboxEventRepositoryPort
+import github.com.hukuta94.delivery.core.application.port.repository.event.InboxEventRepositoryPort
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.model.entity.event.InboxEventJpaEntity
 import org.slf4j.LoggerFactory
